Move seahorse_util to separate file
this makes the generated lib.rs much easier to read
it also sets a pattern for adding future library files as needed
